pyproject.toml
[tool.poetry]
name = "ui-automation-demo"
version = "0.1.0"
description = ""
authors = ["Ash Lam <ash.lam@crypto.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Appium-Python-Client = "^2.8.1"
PyHamcrest = "^2.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
